requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements-dev.in
#
attrs==22.1.0
    # via pytest
iniconfig==1.1.1
    # via pytest
packaging==22.0
    # via pytest
pluggy==1.0.0
    # via pytest
pytest==7.2.0
    # via
    #   -r requirements-dev.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements-dev.in
